+++ packages/app/app-template/app/entities/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { ConfigurationJson } from "@/app/ConfigurationJson";
+
+export default function EntitiesPage() {
+	return <ConfigurationJson title="Entity types" field="entityTypes" />;
+}