+++ packages/app/app-template/app/relationships/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { ConfigurationJson } from "@/app/ConfigurationJson";
+
+export default function RelationshipsPage() {
+	return <ConfigurationJson title="Relation types" field="relationTypes" />;
+}